+++ metatensor-torch/src/atomistic/model.cpp
@@ -140,7 +140,8 @@ std::unordered_set<std::string> KNOWN_OUTPUTS = {
     "energy_ensemble",
     "energy_uncertainty",
     "features",
-    "non_conservative_forces"
+    "non_conservative_forces",
+    "non_conservative_stress"
 };
 
 void ModelCapabilitiesHolder::set_outputs(torch::Dict<std::string, ModelOutput> outputs) {
@@ -1080,6 +1081,12 @@ static std::unordered_map<std::string, Quantity> KNOWN_QUANTITIES = {
     }, {
         // alternative names
         {"eV/A", "eV/Angstrom"},
+    }}},
+    {"stress", Quantity{/* name */ "stress", /* baseline */ "eV/Angstrom^3", {
+        {"eV/Angstrom^3", 1.0},
+    }, {
+        // alternative names
+        {"eV/A^3", "eV/Angstrom^3"},
     }}},
 };